+++ .storybook/stories/TableFiltering/components/FilterBar.tsx
@@ -1,97 +1,46 @@
-import { useSpring, animated } from '@react-spring/web';
-import { Flex } from '@ag.ds-next/react/box';
-import { GetDataFilters, GetDataSort } from '../lib/getData';
-import { DateRangePicker } from '@ag.ds-next/react/date-picker';
-import { Button } from '@ag.ds-next/react/button';
-import { FilterAssigneeSelect } from './FilterAssigneeSelect';
-import { FilterStateSelect } from './FilterStateSelect';
-import { usePrefersReducedMotion } from '@ag.ds-next/react/core';
-import { useRef } from 'react';
+import { Flex, Stack } from '@ag.ds-next/react/box';
+import { ReactNode } from 'react';
 
-export const FilterBar = ({
-	ariaLabelledBy,
-	id,
-	isOpen,
-	filters,
-	setFilters,
-	resetPagination,
-	resetFilters,
-}: {
-	ariaLabelledBy: string;
-	id: string;
-	isOpen: boolean;
-	filters: GetDataFilters;
-	setFilters: (filters: GetDataFilters) => void;
-	resetPagination: () => void;
-	resetFilters: () => void;
-}) => {
-	// This code has been copied from the Accordion component.
-	const ref = useRef<HTMLDivElement>(null);
-	const prefersReducedMotion = usePrefersReducedMotion();
-	const animatedHeight = useSpring({
-		from: { display: 'none', height: 0 },
-		to: async (next) => {
-			if (isOpen) await next({ display: 'block' });
-			await next({
-				height: isOpen ? ref.current?.offsetHeight : 0,
-				immediate: prefersReducedMotion,
-			});
-			await next(isOpen ? { height: 'auto' } : { display: 'none' });
-		},
-	});
-
+export const FilterRegion = ({ children }: { children: ReactNode }) => {
 	return (
-		<animated.section
-			id={id}
-			aria-labelledby={ariaLabelledBy}
-			style={animatedHeight}
-			css={{
-				overflow: 'hidden',
+		<Stack
+			as="section"
+			gap={1}
+			aria-label="data filtering"
+			borderBottom
+			paddingBottom={1}
+		>
+			{children}
+		</Stack>
+	);
+};
+
+export const FilterBar = ({ children }: { children: ReactNode }) => {
+	return (
+		<Flex
+			columnGap={1}
+			rowGap={1.5}
+			alignItems={{ xs: 'stretch', lg: 'flex-end' }}
+			justifyContent={{ xs: 'flex-start', md: 'space-between' }}
+			flexDirection={{
+				xs: 'column-reverse',
+				lg: 'row',
 			}}
 		>
-			<Flex
-				ref={ref}
-				gap={1}
-				background="shade"
-				flexWrap="wrap"
-				alignItems="flex-end"
-				padding={1}
-				rounded
-				border
-				borderColor="muted"
-			>
-				<FilterStateSelect
-					filters={filters}
-					setFilters={setFilters}
-					resetPagination={resetPagination}
-				/>
-				<FilterAssigneeSelect
-					filters={filters}
-					setFilters={setFilters}
-					resetPagination={resetPagination}
-				/>
-				<DateRangePicker
-					fromLabel="Registered from"
-					toLabel="Registered to"
-					onChange={(value) => {
-						resetPagination();
-						setFilters({
-							...filters,
-							requestDate: value,
-						});
-					}}
-					value={filters.requestDate}
-				/>
-				<Button
-					variant="secondary"
-					onClick={() => {
-						resetPagination();
-						resetFilters();
-					}}
-				>
-					Reset filters
-				</Button>
-			</Flex>
-		</animated.section>
+			{children}
+		</Flex>
+	);
+};
+
+export const FilterBarGroup = ({ children }: { children: ReactNode }) => {
+	return (
+		<Flex
+			columnGap={1}
+			rowGap={1.5}
+			flexDirection={{ xs: 'column', md: 'row' }}
+			alignItems={{ xs: 'stretch', md: 'flex-end' }}
+		>
+			{children}
+		</Flex>
 	);
 };
